Restructure to avoid ffmpeg-specific logic in the parsing
Instead of converting/coalescing into the ffmpeg output format during parsing, we parse everything normally and make all the info available in order to coalesce the metadata to the ffmpeg format afterwards. Comparing against ffmpeg's output is just a way of checking our work, not a goal in terms of how parsing/interpreting the parsed data should be done.