Add BufferedStreamSource and use it
This does way less than I was expecting it to. I think I had false hope because I tried naively wrapping the reader in a bufferedReader and saw a significant speedup, but in hindsight that was just because it was (incorrectly) failing early when reading tags since a normal buffered reader will just explode when seeked.