Refactor EncodedTextIterator to try to avoid it being overcomplicated
Instead of having it be a function that returns a type specialized for each encoding, it's just a single struct that uses runtime branching to handle all encodings. This ends up making usage much nicer and handles more use-cases, as it's much easier for the callers to determine exactly how/when the `next` function is called. With the previous method, it had to collect all texts in one function call and used a crazy switch statement and @fieldParentPtr stuff in order to use the right specialized struct.