Revert "When runtime safety is enabled, ensure that we never try to read more than the file size"
This reverts commit 66a360e17b4e63a55d824a431f3d1b6f540dba20.

This doesn't actually find anything worth fixing, as it can trigger on things like trying to read a 10 byte ID3 header in a < 10 byte file (which is totally acceptable, no reason to short-circuit that). The real thing we want to avoid is large allocations, and the custom allocator used in the fuzzer handles that just fine; this reader doesn't actually get us anything new.