mp4: Fix integer overflow on extended size edge case
If the extended size field had a value of > AtomHeader.len but less than AtomHeader.len + AtomHeader.extended_field_len, then it would lead to an overflow when calculating sizeExcludingHeader. Now, we check that the size is big enough to fit the actual full header length, taking into account whether or not an extended size field is present.