collate: Handle Windows-1251 detection correctly
Previously, the UTF-8 bytes were being interpreted as Windows-1251 when doing threshold detection, which was wrong. Instead, the UTF-8 needs to be converted back to extended ASCII before checking the threshold. This commit makes that change while also making sure not to do any extra allocations for doing the UTF-8 -> extended ascii conversion, and instead that is done on-the-fly when detecting/converting Windows-1251.