collate: Make normalizer optional and assert that UTF-8 in the metadata is always valid
After some profiling, I noticed that Normalizer.init/deinit was taking 99% of the runtime. In practice, the normalizer seems likely to only have marginal tangible effects (although this might depend on the language, as some languages have a lot of diacritics that normalization would be useful for), so making it optional makes sense. Allowing a single instance to be shared across multiple Collator instances is also useful to avoid having to pay the cost of Normalizer.init multiple times.