collate: Reinstate Unicode normalization using zg
The initialization cost of the normalization data has been hugely reduced in zg compared to ziglyph, and the runtime performance looks good, so normalization is now the default for Collator.

For case folding, my attempt at an implementation was kept over zg to make usage a bit simpler, since zg wants to pass around FoldData whereas my implementation just uses arrays directly. In benchmarking, zg's implementation is slightly faster (even including initialization cost of the data), but it's close enough for the tradeoff in simplicity to be worth it I think.